refactor: extract image-load state machine + add accessibility
Two follow-ups to the React best-practices review (PR comments #7, #8):

**#8 — Extract `useImageWithRetry` hook.** The fetch state machine
(backoff schedules, abort controllers, timer tracking, ref-stable
listener attachment) was ~200 lines of imperative async logic awkwardly
inlined in a presentation component. Moved to
`src/components/editor/components/blocks/image/useImageWithRetry.ts`,
which returns `{ src, phase, isImageReady, lastError, retry,
onImageLoaded, onImageError }`. `Img.tsx` is now ~100 lines of pure
rendering. The hook is unit-testable in isolation with a fetch mock,
and the JSX no longer hides correctness concerns behind a wall of
useCallback/useEffect plumbing.

**#7 — Accessibility.** Pre-existing issue: every image rendered with
`alt={''}`, which tells assistive tech "this is decorative" — wrong
for content images in a document editor. `Img` now accepts an `alt`
prop. When omitted (current callers), it derives a label from the URL
filename — strips extension, percent-encoding, AppFlowy upload prefix,
collapses separators. Falls back to "Embedded image" if nothing
sensible can be extracted. Future work: thread an explicit `alt` from
ImageBlockData once the data model gains a caption field.

Also tightened the failure/loading overlays with `role="status"`,
`aria-live="polite"`, `aria-label`, and `aria-hidden` on the
decorative error icon — screen readers now announce loading state
transitions instead of going silent.

Behavior unchanged; lint + type-check clean. Image E2E tests not
re-run in this commit (local docker stack down — port 8000 not
listening); they passed against the previous commit on this branch
and the refactor preserves all observable behavior.
